improved logging
