+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.6] -
+
+### Fixed
+
+- Fixed an issue where the api base context was not setting the log correctly
+  resulting in missing log lines
+
 ## [0.4.5] - 2024-01-16
 
 ### Added
